fix #4 Review fixes

+++ calculation/src/test/java/ee/qrental/calculation/application/service/strategy/PartialWeeksRateCalculationStrategyTest.java
@@ -6,6 +6,7 @@ import ee.qrental.transaction.domain.Constant;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.Month;
 
@@ -20,7 +21,6 @@ import static org.mockito.Mockito.when;
  * 2    3    4    5    6    7    8
  */
 
-/*
 class PartialWeeksRateCalculationStrategyTest {
 
     private RateCalculationStrategy instanceUnderTest;
@@ -73,8 +73,6 @@ class PartialWeeksRateCalculationStrategyTest {
         final var result = instanceUnderTest.calculate(week);
 
         //then
-        assertEquals(1000L, result);
+        assertEquals(BigDecimal.valueOf(1000d), result);
     }
 }
-
- */
